# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /audio/tts:
    post:
      operationId: createTts
      tags: [audio]
      summary: Generate a short voiceover preview
      description: |
        Generates an MP3 voiceover preview for a storyboard frame using
        Replit-managed OpenAI audio. Intended for short, founder-authored
        storyboard voiceover lines; long narration exports are out of scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTtsBody"
      responses:
        "200":
          description: MP3 audio bytes
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many voiceover previews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Speech generation failed upstream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /jobs/{jobId}:
    get:
      operationId: getGenerationJob
      tags: [projects]
      summary: Read the current state of a generation job
      description: |
        Returns the latest snapshot of a generation_jobs row so a polling
        client can render progress (current step, done/total counters)
        and react to terminal states (succeeded → refetch project,
        failed → show error_message).
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The current job snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        slug:
          type: string
          description: |
            Optional. Present on POST /projects 500 responses when a project row
            was created before the failure, so the client can deep-link to
            /projects/:slug to view the persisted error run.
      required:
        - error
...
    SpeechVoice:
      type: string
      enum:
        - alloy
        - echo
        - fable
        - onyx
        - nova
        - shimmer
    SpeechFormat:
      type: string
      enum:
        - mp3
    CreateTtsBody:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1000
          description: Short storyboard voiceover text to speak.
        voice:
          $ref: "#/components/schemas/SpeechVoice"
          description: Optional OpenAI voice. Defaults to onyx.
        format:
          $ref: "#/components/schemas/SpeechFormat"
          description: Optional output format. Defaults to mp3.
      required:
        - text
...
    GenerationJobStatus:
      type: string
      enum:
        - queued
        - running
        - succeeded
        - failed
    GenerationJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/GenerationJobStatus"
        progressTotal:
          type: integer
          minimum: 0
          description: |
            Total work units the worker expects to complete. May be 0 until
            the worker has finished the text-generation phase and computed
            the per-image fan-out count.
        progressDone:
          type: integer
          minimum: 0
          description: |
            Work units completed so far. Reaches `progressTotal` when the
            job hits a terminal state (`succeeded`).
        currentStep:
          oneOf:
            - type: string
            - type: "null"
          description: |
            Human-readable label for the worker's current phase (e.g.
            "Generating launch copy", "Generating image 3/11", "Done").
        errorMessage:
          oneOf:
            - type: string
            - type: "null"
          description: |
            Set when status is `failed`. Clamped server-side to 4000 chars
            so a multi-MB upstream stack trace can never blow up the row.
        startedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        completedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - status
        - progressTotal
        - progressDone
        - currentStep
        - errorMessage
        - startedAt
        - completedAt
        - createdAt
        - updatedAt
...